fix: la-132 rename someAuth file
