+++ src/middleware/validateTokenAndUrlWithInvitedUser.ts
